_zip_extract/savdoai-main/requirements.txt
Pillow>=10.0.0
anthropic>=0.49.0
asyncpg==0.29.0
beautifulsoup4>=4.12.0
celery[redis]>=5.3.0
ebooklib>=0.18
fastapi>=0.110.0
google-genai>=1.10.0
httpx>=0.28.1
lxml>=5.0.0
openpyxl==3.1.5
pdfplumber>=0.10.0
pydantic>=2.0.0
pydub>=0.25.1
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx>=0.6.23
python-telegram-bot[job-queue]==21.3
pytz==2024.1
redis[hiredis]>=5.0.0
reportlab==4.2.5
striprtf>=0.0.26
tenacity>=8.2.0
uvicorn[standard]>=0.29.0
websockets>=12.0
qrcode[pil]>=7.4
